+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2890 Add GHS hazard categories to SampleType and Sample
+- #2881 Add text support for interim fields
 - #2888 Improve ID Server admin views
 - #2889 Remove orphan temp objects when renameAfterCreation fails
 - #2887 Set default Seq Type to 'generated' for partition/retest/secondary